test: arpa_inet conversion functions tests
