+++ src/compat/posix/net/tests/Mybuild
@@ -1,7 +1,7 @@
 package embox.test.net
 
-module inet_pton_test {
-	source "inet_pton_test.c"
+module arpa_inet_test {
+	source "arpa_inet_test.c"
 
 	depends embox.framework.test
 }
