+++ src/compat/posix/net/tests/arpa_inet_test.c
@@ -10,11 +10,24 @@
 #include <framework/test/assert.h>
 #include <kernel/printk.h>
 
-#include "inet_pton_test.h"
+#include "arpa_inet_test.h"
 
-EMBOX_TEST_SUITE("inet_pton behaviour testsuite");
+EMBOX_TEST_SUITE("inet_pton and inet_aton behaviour testsuite");
 
-TEST_CASE("inet_pton af=AF_INET") {
+TEST_CASE("embox inet_aton results against linux version") {
+	int res;
+	struct in_addr v4_in;
+
+	for (int i = 0; i < V4_TESTS; i++) {
+		v4_in.s_addr = 0;
+		res = inet_aton(v4_test_ref[i].test_string, &v4_in);
+		test_assert_equal(res, v4_test_ref[i].aton_pass);
+		test_assert_equal(v4_in.s_addr, v4_test_ref[i].aton_val.s_addr);
+        printk(".");
+	}
+}
+
+TEST_CASE("inet_pton af=AF_INET results against linux version") {
 	int af = AF_INET;
 	int res;
 	struct in_addr v4_in;
@@ -22,13 +35,13 @@ TEST_CASE("inet_pton af=AF_INET") {
 	for (int i = 0; i < V4_TESTS; i++) {
 		v4_in.s_addr = 0;
 		res = inet_pton(af, v4_test_ref[i].test_string, &v4_in);
-		test_assert_equal(res, v4_test_ref[i].pass);
-		test_assert_equal(v4_in.s_addr, v4_test_ref[i].true_val.s_addr);
-		printk(".");
+		test_assert_equal(res, v4_test_ref[i].pton_pass);
+		test_assert_equal(v4_in.s_addr, v4_test_ref[i].pton_val.s_addr);
+        printk(".");
 	}
 }
 
-TEST_CASE("inet_pton af=AF_INET6") {
+TEST_CASE("inet_pton af=AF_INET6 results against linux version") {
 	int af = AF_INET6;
 	int res;
 	struct in6_addr v6_in;
